requirements.txt
google-cloud-documentai
pandas
openpyxl
reportlab